[KAIZEN-0] rydder i less og struktur og gjør om til ekspanderbart panel
