+++ package.json
@@ -9,6 +9,7 @@
     "lodash.throttle": "^4.1.1",
     "nav-frontend-alertstriper": "^1.0.8",
     "nav-frontend-alertstriper-style": "^0.2.33",
+    "nav-frontend-chevron": "^1.0.5",
     "nav-frontend-chevron-style": "^0.3.4",
     "nav-frontend-core": "^4.0.5",
     "nav-frontend-ekspanderbartpanel": "^1.0.3",
@@ -17,6 +18,8 @@
     "nav-frontend-js-utils": "^1.0.6",
     "nav-frontend-knapper": "^1.0.10",
     "nav-frontend-knapper-style": "^0.3.9",
+    "nav-frontend-lenker": "^1.0.8",
+    "nav-frontend-lenker-style": "^0.2.6",
     "nav-frontend-paneler-style": "^0.3.7",
     "nav-frontend-spinner": "^1.0.8",
     "nav-frontend-spinner-style": "^0.2.4",
