+++ src/app/basisinfo/basisinfo.less
@@ -1,0 +1,4 @@
+.basisinfo__ikon {
+  height: 2rem;
+  width: 2rem;
+}
