+++ src/app/persondetaljer.tsx
@@ -2,11 +2,10 @@ import * as cls from 'classnames';
 import * as React from 'react';
 
 import Basisinfo from "./basisinfo/basisinfo";
-import Informasjonsvisning from "./informasjonsvisning/informasjonsvisning";
-import Knappelinje from './knappelinje/knappelinje';
 import Tilbakelenke from "./tilbakelenke";
+import Informasjonsvisning from "./visningskomponenter/informasjonsvisning";
 
-import { AppContext, IAppContextProp, withAppContext } from "../context";
+import { AppContext, IAppContextProp, withAppContext } from "./context";
 
 import './persondetaljer.less';
 
@@ -18,17 +17,13 @@ class Persondetaljer extends React.Component<IAppContextProp> {
             <>
                 <Tilbakelenke />
                 <div
-                    className={cls("panel--stor ", {
+                    className={cls("persondetaljer", {
                         'apen': apen,
                         'lukket': !apen
                     })}
                 >
-                    <div className="persondetaljer">
-                        <Basisinfo />
-
-                        <Knappelinje />
-                        <Informasjonsvisning />
-                    </div>
+                    <Basisinfo />
+                    <Informasjonsvisning />
                 </div>
             </>
         )
