--- src/app/persondetaljer/knappelinje/knappelinje.less
@@ -1,34 +1,0 @@
-@import "~nav-frontend-core/less/_variabler";
-@import "~nav-frontend-typografi-style/src/fontimports";
-@import "~nav-frontend-knapper-style/src/mixins";
-
-.knappelinje {
-  .knappelinje__toggleknapp {
-    .knappbase-mixin(@font-family);
-
-    padding: 0.25rem 0.75rem;
-    background-color: @navLysGra;
-    color: @navMorkGra;
-    letter-spacing: normal;
-    border: none;
-    text-transform: none;
-    text-decoration: none;
-    line-height: inherit;
-
-    margin-right: 0.5rem;
-
-    &:hover {
-      color: @navBla;
-      text-decoration: underline;
-    }
-  }
-  .knappelinje__toggleknapp--valgt {
-      background-color: @navBla;
-      color: white;
-
-    &:hover {
-      background-color: @navDypBla;
-      color: white;
-    }
-  }
-}
