--- src/app/persondetaljer/tilbakelenke.tsx
@@ -1,9 +1,0 @@
-import * as React from 'react';
-
-const Tilbakelenke = () =>
-    <a href={`/veilarbportefoljeflatefs/tilbake?enhet=${1234 /* valgtEnhet */}`}> {/* TODO: hente enhet */}
-        <i className="chevron--venstre"/>
-        <span>Tilbake</span> {/* Todo: tekst */}
-    </a>;
-
-export default Tilbakelenke;
