+++ src/app/utils/datafetcher.tsx
@@ -40,5 +40,4 @@ class Datafetcher<T> extends React.Component<IProps<T>, IState<T>> {
     }
 }
 
-
 export default Datafetcher;
