+++ src/app/visningskomponenter/informasjonsvisning.less
@@ -1,4 +1,8 @@
-@import "~nav-frontend-core/less/_variabler";
+@import "../../../node_modules/nav-frontend-core/less/_variabler";
+
+.informasjonsvisning .ekspanderbartPanel {
+  margin-bottom: 4px;
+}
 
 .informasjonsbolk {
   margin: 1rem 0 1rem 0;
